+++ src/decoder-opennmt/src/main/java/eu/modernmt/decoder/opennmt/memory/ScoreEntry.java
@@ -5,13 +5,13 @@ package eu.modernmt.decoder.opennmt.memory;
  */
 public class ScoreEntry implements Comparable<ScoreEntry> {
 
-    public final int domain;
+    public final long domain;
     public final String[] sentence;
     public final String[] translation;
 
     public float score = 0.f;
 
-    public ScoreEntry(int domain, String[] sentence, String[] translation) {
+    public ScoreEntry(long domain, String[] sentence, String[] translation) {
         this.domain = domain;
         this.sentence = sentence;
         this.translation = translation;
